GridNoSpotterCTHBonus: fixed bTeam type.
git-svn-id: https://ja2svn.mooo.com/source/ja2/trunk/GameSource/ja2_v1.13/Build@8949 3b4a5df2-a311-0410-b5c6-a8a6f20db521

+++ Tactical/Soldier Control.cpp
@@ -25094,7 +25094,7 @@ BOOLEAN PlayerTeamIsScanning( )
 }
 
 // bonus for snipers firing at this location (we get this if there are spotters)
-UINT16	GridNoSpotterCTHBonus( SOLDIERTYPE* pSniper, INT32 sGridNo, UINT bTeam )
+UINT16	GridNoSpotterCTHBonus( SOLDIERTYPE* pSniper, INT32 sGridNo, INT8 bTeam )
 {
 	UINT16 bestvalue = 0;
